pyproject.toml
[tool.poetry]
name = "quizapp"
version = "0.1.0"
description = ""
authors = ["Saajan <saajan.n.joshi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
